Milestone Capture
We can’t validate venue in the Performance model because venue data is
not reaching the model. When the validation is forced, the views throw
an error (as long as venue is requested in the views). So we have to
not call for venue in the views. But, venue data is still not collected
in the db (unless manually thru Admin). We will try to bypass the
checkin model by adding a boolean checked_in column in the performance
table. So instead of building a checkin in row, we’ll use the
performance table to account for this. “Check In” will start in the
Performance table.   We can “fake” it, by still being able to access
venue in views, but it still will not be captured by the database,
which may cause unintended consequences.