Adds Singular Request Functionality
Users can now only make one request at a time. This is a successful
workaround to the previous issue.
Users can be removed from the request queue in an instance where they
want to request a different song, but cannot because they have an
existing in the queue. This solution most closely mimics real life.